Some significant changes
Introduced EventRouter and decoupled AggregateRoot
from AggregateState

Started working more seriously on application services
and repository implementations, hopefully I'll have more
time.

+++ src/EventSourcing.Example/ExampleMessages.cs
@@ -36,6 +36,6 @@ namespace EventSourcing.Example
         public ExampleId Id { get; set; }
 
         [DataMember(Order = 2)]
-        public long ExpectedVersion { get; set; }
+        public int ExpectedVersion { get; set; }
     }
 }
